test_requirements.txt
black==20.8b1
bpython
codecov
ddt
django-debug-toolbar
factory_boy
faker
ipdb
isort
moto
nplusone
pdbpp
pip-tools
pylint==2.6.0
pylint-django==2.3.0
pytest
pytest-cov
pytest-env==0.6.2
pytest-django
pytest-env==0.6.2
pytest-mock
semantic-version
